backend: Email: create our own mjml-react lib (#29428)
Currently we're using `mjml-react` to create react nodes.

This code replaces the parts of the library we use in a lightweight and maintainable way. This will give us the flexibility needed to modify the library to fit our needs.

- drop in replacement (no code changes needed to update)
- better types
- generated so custom components can be added
- unlocks ability to add features like React Context and `className` -> `css-class`

In order to merge styled-components (#28399) we need Mjml components that take a `className`.

React Context of which MJML component a component is being rendered inside
React context would hold a stack of MJML components.
Our components can decided how to render based on which wrapping MJML component is being used.
Currently this is done manually using the `raw` tags (inflexible, devs have to remember to use it).